// constants/types/homeType.ts
export type SizeImage = {
  width?: string;
  height?: string;
  display?: 'flex';
};

export type ProductCar = {  
  id: number;
  name: string;
  email: string;
  gender: string;
};

export type ProductDetail = {
  id: string;
  name: string;
  description: string;
  image: string;
  price: string;
  fuel: string;
  power: string;
};